+++ HISTORY.md
@@ -3,6 +3,7 @@
 # 0.6.0
 * ENH download metadata template in the GUI
 * ENH multicova analysis
+* ENH filter data completeness `dataset.preprocess(data_completeness=0.7)` 
 
 # 0.5.4
 * FIX altair version - binning of streamlit version